fix layout and functionalities
