--- src/components/Footer/Footer.js
@@ -1,23 +1,0 @@
-import styled from "styled-components";
-export const Container = styled.div`
-     width: 90%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 20px;
-    overflow-y: scroll;
-    scrollbar-width: none;
-    h1{
-        font-size: 15px;
-        border-bottom: 1px solid #D3D3D3;
-    }
-    h2{
-        font-size: 10px;
-        color:#D3D3D3;
-    }
-    Link{
-        text-decoration: none;
-    }
-`
-export const Institucional = styled.div``
-export const Acount = styled.div``
