+++ src/components/Logo/Logo.js
@@ -4,6 +4,7 @@ export const Container = styled.div`
     display: flex;
     justify-content: space-between;
     align-items: center;
+    
 `
 export const First = styled.div`
     font-family: 'Bitter', serif;
@@ -14,4 +15,9 @@ export const Second = styled.div`
     font-family: 'Molle', cursive;
     color:#B23428;
     font-size: 60px;
+`
+export const Inside = styled.div`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
 `
