+++ src/components/Logo/Logo.jsx
@@ -1,9 +1,14 @@
-import {Container, First, Second} from "./Logo"
+import {Container, First, Second, Inside} from "./Logo"
+import { Link } from "react-router-dom"
 export default function Logo(){
     return(
         <Container>
-             <First>Empório</First> 
-             <Second>Driven</Second>
+            <Link to={`/`}>
+                <Inside>
+                    <First>Empório</First> 
+                    <Second>Driven</Second>
+                </Inside>
+            </Link>
         </Container>
     )
 }
