+++ src/components/ProductPage/ProductPage.jsx
@@ -3,7 +3,6 @@ import { Link } from "react-router-dom";
 import { useState, useEffect, useContext} from "react";
 import axios from "axios";
 import { useParams } from "react-router-dom";
-import { Oval } from "react-loader-spinner";
 import CartContext from "../../contexts/CartContext.js";
 
 export default function ProductPage(){
@@ -65,40 +64,6 @@ export default function ProductPage(){
                   </Buy>
               </Info>
              </Product>  
-          <Related>
-              <h3>Produtos Relacionados</h3>
-             
-              <RelatedProduct>
-              <ProductList list={categoryRender} />
-                
-              </RelatedProduct>
-          </Related>
         </Container>
     )
-}
-function ProductList({ list, index }) {
-  if (list.length > 0) {
-    return (
-      <>
-        {list.map((product) => {
-          return (
-            <>
-              <EachProduct key ={index}>
-              <Image>
-                  <img src={product.image} alt="imagem do produto"/>
-              </Image>
-              <h4>{product.name}</h4>
-              <h5>R${product.price}</h5>
-              <ButtonRelated>Comprar</ButtonRelated>
-            </EachProduct>
-
-         
-            </>
-         );
-        })}
-      </>
-    );
-  } else {
-    return <Oval color="#ff8b1e" height={80} width={80} />;
-  }
 }
